Revert "fix warning: variable X might be clobbered by 'longjmp'"
This reverts commit b092b63119cbfe3cb4bc786eee81630998996acf which
introduced a wrapper function around setjmp(). To quote from man setjmp:

  If the function which called setjmp() returns before longjmp() is
  called, the behavior is undefined.  Some kind of subtle or unsubtle
  chaos is sure to result.

Since after the above commit setjmp() is called from the wrapper
function, the result might or might not work, depending on compiler
settings. If the setjmp() wrapper is not inlined, then the state of the
stack after longjmp() will likely be garbage.

[Cherry-Pick:]

This is a cherry-pick from a master branch to the 1.14 branch. The
reverted commit was cherry-picked to this branch in commit
d53db01d01a48c48a1633a8d531f979a99d316bd, so this is what this commit
actually reverts.