Hide skia backend from users
Comment out the skia backend in configure.ac to avoid presenting it as
an option to users.

Skia is not API stable and is not available in a packaged/versioned
form.  This results in Cairo's skia backend constantly being out of date
and broken, since it will only work with a specific Skia snapshot.  The
last time the cairo skia backend was updated was 2014, and it hadn't
been updated very regularly prior to that.  For these reasons, unless
someone raises an interest in using and/or maintaining the backend,
we'll consider dropping it entirely.